pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gsheets-sql"
version = "0.1.0"
description = "Use Google Sheets like a PostgreSQL database"
readme = "specs.md"
requires-python = ">=3.10"
dependencies = [
    "gspread>=6.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "pandas>=2.0",
    "duckdb>=0.10",
    "sqlglot>=23.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
]

[project.scripts]
gsheets-sql = "gsheets_sql.repl:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["gsheets_sql*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
